Revert PaymentSheet+API changes, move to PaymentSheetLoader
