+++ Stripe/StripeiOSTests/STPIntentWithPreferencesTest.swift
@@ -21,139 +21,120 @@ class STPIntentWithPreferencesTest: XCTestCase {
     private let setupIntentClientSecret =
         "seti_1GGCuIFY0qyl6XeWVfbQK6b3_secret_GnoX2tzX2JpvxsrcykRSVna2lrYLKew"
 
-    func testPaymentIntentWithPreferences() {
-        let expectation = XCTestExpectation(description: "Retrieve Payment Intent With Preferences")
+    func testPaymentIntentWithPreferences() async {
         let client = STPAPIClient(publishableKey: STPTestingDefaultPublishableKey)
 
-        client.retrievePaymentIntentWithPreferences(withClientSecret: paymentIntentClientSecret) {
-            result in
-            switch result {
-            case .success(let paymentIntentWithPreferences):
-                expectation.fulfill()
-                // Check for required PI fields
-                XCTAssertEqual(paymentIntentWithPreferences.stripeId, "pi_1H5J4RFY0qyl6XeWFTpgue7g")
-                XCTAssertEqual(
-                    paymentIntentWithPreferences.clientSecret,
-                    self.paymentIntentClientSecret
-                )
-                XCTAssertEqual(paymentIntentWithPreferences.amount, 2000)
-                XCTAssertEqual(paymentIntentWithPreferences.currency, "usd")
-                XCTAssertEqual(
-                    paymentIntentWithPreferences.status,
-                    STPPaymentIntentStatus.succeeded
-                )
-                XCTAssertEqual(paymentIntentWithPreferences.livemode, false)
-                XCTAssertEqual(
-                    paymentIntentWithPreferences.paymentMethodTypes,
-                    STPPaymentMethod.types(from: ["card"])
-                )
-                // Check for ordered payment method types
-                XCTAssertNotNil(paymentIntentWithPreferences.orderedPaymentMethodTypes)
-                XCTAssertEqual(
-                    paymentIntentWithPreferences.orderedPaymentMethodTypes,
-                    [STPPaymentMethodType.card]
-                )
-            case .failure(let error):
-                print(error)
-            }
+        do {
+            let paymentIntentWithPreferences = try await client.retrievePaymentIntentWithPreferences(withClientSecret: paymentIntentClientSecret)
+            // Check for required PI fields
+            XCTAssertEqual(paymentIntentWithPreferences.stripeId, "pi_1H5J4RFY0qyl6XeWFTpgue7g")
+            XCTAssertEqual(
+                paymentIntentWithPreferences.clientSecret,
+                self.paymentIntentClientSecret
+            )
+            XCTAssertEqual(paymentIntentWithPreferences.amount, 2000)
+            XCTAssertEqual(paymentIntentWithPreferences.currency, "usd")
+            XCTAssertEqual(
+                paymentIntentWithPreferences.status,
+                STPPaymentIntentStatus.succeeded
+            )
+            XCTAssertEqual(paymentIntentWithPreferences.livemode, false)
+            XCTAssertEqual(
+                paymentIntentWithPreferences.paymentMethodTypes,
+                STPPaymentMethod.types(from: ["card"])
+            )
+            // Check for ordered payment method types
+            XCTAssertNotNil(paymentIntentWithPreferences.orderedPaymentMethodTypes)
+            XCTAssertEqual(
+                paymentIntentWithPreferences.orderedPaymentMethodTypes,
+                [STPPaymentMethodType.card]
+            )
+        } catch {
+            XCTFail()
+            print(error)
         }
-        wait(for: [expectation], timeout: STPTestingNetworkRequestTimeout)
     }
 
-    func testSetupIntentWithPreferences() {
-        let expectation = XCTestExpectation(description: "Retrieve Setup Intent With Preferences")
+    func testSetupIntentWithPreferences() async {
         let client = STPAPIClient(publishableKey: STPTestingDefaultPublishableKey)
 
-        client.retrieveSetupIntentWithPreferences(withClientSecret: setupIntentClientSecret) {
-            result in
-            switch result {
-            case .success(let setupIntentWithPreferences):
-                expectation.fulfill()
-                // Check required SI fields
-                XCTAssertEqual(setupIntentWithPreferences.stripeID, "seti_1GGCuIFY0qyl6XeWVfbQK6b3")
-                XCTAssertEqual(
-                    setupIntentWithPreferences.clientSecret,
-                    self.setupIntentClientSecret
-                )
-                XCTAssertEqual(setupIntentWithPreferences.status, .requiresPaymentMethod)
-                XCTAssertEqual(
-                    setupIntentWithPreferences.paymentMethodTypes,
-                    STPPaymentMethod.types(from: ["card"])
-                )
-                // Check for ordered payment method types
-                XCTAssertNotNil(setupIntentWithPreferences.orderedPaymentMethodTypes)
-                XCTAssertEqual(
-                    setupIntentWithPreferences.orderedPaymentMethodTypes,
-                    [STPPaymentMethodType.card]
-                )
-            case .failure(let error):
-                print(error)
-            }
+        do {
+            let setupIntentWithPreferences = try await client.retrieveSetupIntentWithPreferences(withClientSecret: setupIntentClientSecret)
+            // Check required SI fields
+            XCTAssertEqual(setupIntentWithPreferences.stripeID, "seti_1GGCuIFY0qyl6XeWVfbQK6b3")
+            XCTAssertEqual(
+                setupIntentWithPreferences.clientSecret,
+                self.setupIntentClientSecret
+            )
+            XCTAssertEqual(setupIntentWithPreferences.status, .requiresPaymentMethod)
+            XCTAssertEqual(
+                setupIntentWithPreferences.paymentMethodTypes,
+                STPPaymentMethod.types(from: ["card"])
+            )
+            // Check for ordered payment method types
+            XCTAssertNotNil(setupIntentWithPreferences.orderedPaymentMethodTypes)
+            XCTAssertEqual(
+                setupIntentWithPreferences.orderedPaymentMethodTypes,
+                [STPPaymentMethodType.card]
+            )
+        } catch {
+            print(error)
+            XCTFail()
         }
-        wait(for: [expectation], timeout: STPTestingNetworkRequestTimeout)
     }
 
-    func testRetrieveElementSession_deferredPayment() {
+    func testRetrieveElementSession_deferredPayment() async {
         let expectation = XCTestExpectation(description: "Retrieve ElementsSession")
         let client = STPAPIClient(publishableKey: STPTestingDefaultPublishableKey)
-
+        
         let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 2000,
                                                                            currency: "USD",
                                                                            setupFutureUsage: .onSession),
                                                             paymentMethodTypes: ["card", "cashapp"],
                                                             confirmHandler: { _, _, _ in })
-
-        client.retrieveElementsSession(withIntentConfig: intentConfig) { result in
-            switch result {
-            case .success(let elementsSession):
-                XCTAssertNotNil(elementsSession)
-                XCTAssertEqual(elementsSession.countryCode, "US")
-                XCTAssertNotNil(elementsSession.linkSettings)
-                XCTAssertNotNil(elementsSession.paymentMethodSpecs)
-                XCTAssertEqual(
-                    elementsSession.orderedPaymentMethodTypes,
-                    [STPPaymentMethodType.card, STPPaymentMethodType.cashApp]
-                )
-
-                expectation.fulfill()
-            case .failure(let error):
-                print(error)
-            }
+        
+        do {
+            let elementsSession = try await client.retrieveElementsSession(withIntentConfig: intentConfig)
+            XCTAssertNotNil(elementsSession)
+            XCTAssertEqual(elementsSession.countryCode, "US")
+            XCTAssertNotNil(elementsSession.linkSettings)
+            XCTAssertNotNil(elementsSession.paymentMethodSpecs)
+            XCTAssertEqual(
+                elementsSession.orderedPaymentMethodTypes,
+                [STPPaymentMethodType.card, STPPaymentMethodType.cashApp]
+            )
+        } catch {
+            print(error)
+            XCTFail()
         }
-        wait(for: [expectation], timeout: STPTestingNetworkRequestTimeout)
     }
 
-    func testRetrieveElementSession_deferredSetup() {
-        let expectation = XCTestExpectation(description: "Retrieve ElementsSession")
+    func testRetrieveElementSession_deferredSetup() async {
         let client = STPAPIClient(publishableKey: STPTestingDefaultPublishableKey)
 
         let intentConfig = PaymentSheet.IntentConfiguration(mode: .setup(currency: "USD",
                                                                            setupFutureUsage: .offSession),
                                                             paymentMethodTypes: ["card", "cashapp"],
                                                             confirmHandler: { _, _, _ in })
-
-        client.retrieveElementsSession(withIntentConfig: intentConfig) { result in
-            switch result {
-            case .success(let elementsSession):
-                XCTAssertNotNil(elementsSession)
-                XCTAssertEqual(elementsSession.countryCode, "US")
-                XCTAssertNotNil(elementsSession.linkSettings)
-                XCTAssertNotNil(elementsSession.paymentMethodSpecs)
-                XCTAssertEqual(
-                    elementsSession.orderedPaymentMethodTypes,
-                    [STPPaymentMethodType.card, STPPaymentMethodType.cashApp]
-                )
-
-                expectation.fulfill()
-            case .failure(let error):
-                print(error)
-            }
+        
+        do {
+            let elementsSession = try await client.retrieveElementsSession(withIntentConfig: intentConfig)
+            XCTAssertNotNil(elementsSession)
+            XCTAssertEqual(elementsSession.countryCode, "US")
+            XCTAssertNotNil(elementsSession.linkSettings)
+            XCTAssertNotNil(elementsSession.paymentMethodSpecs)
+            XCTAssertEqual(
+                elementsSession.orderedPaymentMethodTypes,
+                [STPPaymentMethodType.card, STPPaymentMethodType.cashApp]
+            )
+        } catch {
+            print(error)
+            XCTFail()
         }
-        wait(for: [expectation], timeout: STPTestingNetworkRequestTimeout)
     }
-
+    
     // MARK: PaymentSheet.IntentConfiguration+elementsSessionPayload tests
-
+    
     func testElementsSessionPayload_Payment() throws {
         let intentConfig = PaymentSheet.IntentConfiguration(mode: .payment(amount: 2000,
                                                                            currency: "USD",
@@ -163,7 +144,7 @@ class STPIntentWithPreferencesTest: XCTestCase {
                                                             onBehalfOf: "acct_connect",
                                                             confirmHandler: { _, _, _ in })
 
-        let payload = intentConfig.elementsSessionPayload(publishableKey: "pk_test")
+        let payload = intentConfig.elementsSessionParameters(publishableKey: "pk_test")
         XCTAssertEqual(payload["key"] as? String, "pk_test")
         XCTAssertEqual(payload["locale"] as? String, Locale.current.toLanguageTag())
 
@@ -184,7 +165,7 @@ class STPIntentWithPreferencesTest: XCTestCase {
                                                             onBehalfOf: "acct_connect",
                                                             confirmHandler: { _, _, _ in })
 
-        let payload = intentConfig.elementsSessionPayload(publishableKey: "pk_test")
+        let payload = intentConfig.elementsSessionParameters(publishableKey: "pk_test")
         XCTAssertEqual(payload["key"] as? String, "pk_test")
         XCTAssertEqual(payload["locale"] as? String, Locale.current.toLanguageTag())
