+++ Stripe/StripeiOSTests/STPTestingAPIClient+Swift.swift
@@ -1,0 +1,78 @@
+//
+//  STPTestingAPIClient+Swift.swift
+//  StripeiOSTests
+//
+//  Created by Yuki Tokuhiro on 6/25/23.
+//
+
+import Foundation
+extension STPTestingAPIClient {
+    static var shared: STPTestingAPIClient {
+        return .shared()
+    }
+    
+    func fetchPaymentIntent(
+        types: [String],
+        currency: String = "eur",
+        paymentMethodID: String? = nil,
+        confirm: Bool = false,
+        completion: @escaping (Result<(String), Error>) -> Void
+    ) {
+        var params = [String: Any]()
+        params["amount"] = 1050
+        params["currency"] = currency
+        params["payment_method_types"] = types
+        params["confirm"] = confirm
+        if let paymentMethodID = paymentMethodID {
+            params["payment_method"] = paymentMethodID
+        }
+
+        createPaymentIntent(
+            withParams: params
+        ) { clientSecret, error in
+            guard let clientSecret = clientSecret,
+                  error == nil
+            else {
+                completion(.failure(error!))
+                return
+            }
+            
+            completion(.success(clientSecret))
+        }
+    }
+    
+    func fetchPaymentIntent(
+        types: [String],
+        currency: String = "eur",
+        paymentMethodID: String? = nil,
+        confirm: Bool = false
+    ) async throws -> String {
+        try await withCheckedThrowingContinuation { continuation in
+            fetchPaymentIntent(
+                types: types,
+                currency: currency,
+                paymentMethodID: paymentMethodID,
+                confirm: confirm
+            ) { result in
+                continuation.resume(with: result)
+            }
+        }
+    }
+
+    func fetchSetupIntent(types: [String], completion: @escaping (Result<(String), Error>) -> Void) {
+        createSetupIntent(
+            withParams: [
+                "payment_method_types": types,
+            ]
+        ) { clientSecret, error in
+            guard let clientSecret = clientSecret,
+                  error == nil
+            else {
+                completion(.failure(error!))
+                return
+            }
+            
+            completion(.success(clientSecret))
+        }
+    }
+}
