+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Link/LinkPaymentController.swift
@@ -77,7 +77,7 @@ import UIKit
     @MainActor
     @_spi(LinkOnly) public func present(from presentingViewController: UIViewController) async throws {
         let linkController: PayWithLinkWebController = try await withCheckedThrowingContinuation { [self] continuation in
-            PaymentSheet.load(mode: mode, configuration: configuration) { result in
+            PaymentSheetLoader.load(mode: mode, configuration: configuration) { result in
                 switch result {
                 case .success(let intent, _, let isLinkEnabled):
                     guard isLinkEnabled else {
@@ -134,7 +134,7 @@ import UIKit
         if (intent == nil || paymentOption == nil) && LinkAccountService().hasSessionCookie {
             // If the customer has a Link cookie, `present` may not need to have been called - try to load here
             paymentOption = try await withCheckedThrowingContinuation { [self] continuation in
-                PaymentSheet.load(mode: mode, configuration: configuration) { result in
+                PaymentSheetLoader.load(mode: mode, configuration: configuration) { result in
                     switch result {
                     case .success(let intent, _, let isLinkEnabled):
                         guard isLinkEnabled else {
