+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet+API.swift
@@ -16,15 +16,7 @@ import UIKit
 @available(iOSApplicationExtension, unavailable)
 @available(macCatalystApplicationExtension, unavailable)
 extension PaymentSheet {
-    /// `PaymentSheet.load()` result.
-    enum LoadingResult {
-        case success(
-            intent: Intent,
-            savedPaymentMethods: [STPPaymentMethod],
-            isLinkEnabled: Bool
-        )
-        case failure(Error)
-    }
+
 
     /// Confirms a PaymentIntent with the given PaymentOption and returns a PaymentResult
     static func confirm(
@@ -335,330 +327,7 @@ extension PaymentSheet {
         }
     }
 
-    /// Fetches the PaymentIntent or SetupIntent and Customer's saved PaymentMethods
-    static func load(
-        mode: InitializationMode,
-        configuration: Configuration,
-        completion: @escaping (LoadingResult) -> Void
-    ) {
-        let intentPromise = Promise<Intent>()
-        let paymentMethodsPromise = Promise<[STPPaymentMethod]>()
-        let loadSpecsPromise = Promise<Void>()
-
-        intentPromise.observe { result in
-            switch result {
-            case .success(let intent):
-                paymentMethodsPromise.observe { result in
-                    switch result {
-                    case .success(let paymentMethods):
-                        // Filter out payment methods that the PI/SI or PaymentSheet doesn't support
-                        // TODO: We're fetching the customer's saved card and us_bank_account PMs, and then filtering - this is backwards!
-                        let savedPaymentMethods = paymentMethods
-                            .filter { intent.recommendedPaymentMethodTypes.contains($0.type) }
-                            .filter {
-                                $0.paymentSheetPaymentMethodType().supportsSavedPaymentMethod(
-                                    configuration: configuration,
-                                    intent: intent
-                                )
-                            }
-
-                        // Warn the merchant if we see unactivated payment method types in the Intent
-                        warnUnactivatedIfNeeded(unactivatedPaymentMethodTypes: intent.unactivatedPaymentMethodTypes)
-                        // Ensure that there's at least 1 payment method type available for the intent and configuration.
-                        let paymentMethodTypes = PaymentMethodType.filteredPaymentMethodTypes(from: intent, configuration: configuration)
-                        guard !paymentMethodTypes.isEmpty else {
-                            completion(.failure(PaymentSheetError.noPaymentMethodTypesAvailable(intentPaymentMethods: intent.recommendedPaymentMethodTypes)))
-                            return
-                        }
-
-                        let linkAccountPromise = PaymentSheet.lookupLinkAccount(
-                            intent: intent,
-                            configuration: configuration
-                        )
-
-                        loadSpecsPromise.observe { _ in
-                            // Overwrite the form specs that were already loaded from disk
-//                            switch intent {
-//                            case .paymentIntent(let paymentIntent):
-//                                _ = FormSpecProvider.shared.loadFrom(paymentIntent.allResponseFields["payment_method_specs"] ?? [String: Any]())
-//                            case .setupIntent:
-//                                break // Not supported
-//                            case .deferredIntent(elementsSession: let elementsSession, intentConfig: _):
-//                                _ = FormSpecProvider.shared.loadFrom(elementsSession.paymentMethodSpecs as Any)
-//                            }
-//                            linkAccountPromise.observe { linkAccountResult in
-//                                switch linkAccountResult {
-//                                case .success(let linkAccount):
-//                                    LinkAccountContext.shared.account = linkAccount
-//
-//                                    completion(
-//                                        .success(
-//                                            intent: intent,
-//                                            savedPaymentMethods: savedPaymentMethods,
-//                                            isLinkEnabled: intent.supportsLink
-//                                        )
-//                                    )
-//                                case .failure:
-//                                    LinkAccountContext.shared.account = nil
-//
-//                                    // Move forward without Link
-//                                    completion(
-//                                        .success(
-//                                            intent: intent,
-//                                            savedPaymentMethods: savedPaymentMethods,
-//                                            isLinkEnabled: false
-//                                        )
-//                                    )
-//                                }
-//                            }
-                        }
-                    case .failure(let error):
-                        completion(.failure(error))
-                    }
-                }
-            case .failure(let error):
-                completion(.failure(error))
-            }
-        }
-        
-        Task {
-            do {
-                // 1a. Fetch PaymentIntent, SetupIntent, or ElementsSession
-                async let _intent = fetchIntent(mode: mode, configuration: configuration)
-                
-                // 1b. List the Customer's saved PaymentMethods
-                async let savedPaymentMethods = fetchSavedPaymentMethods(configuration: configuration)
-
-                // 1c. Load misc singletons
-                async let _ = loadMiscellaneousSingletons() // note: Swift implicitly waits for this before exiting the scope.
-                
-                let intent = try await _intent
-                // Overwrite the form specs that were already loaded from disk
-                switch intent {
-                case .paymentIntent(let paymentIntent):
-                    _ = FormSpecProvider.shared.loadFrom(paymentIntent.allResponseFields["payment_method_specs"] ?? [String: Any]())
-                case .setupIntent:
-                    break // Not supported
-                case .deferredIntent(elementsSession: let elementsSession, intentConfig: _):
-                    _ = FormSpecProvider.shared.loadFrom(elementsSession.paymentMethodSpecs as Any)
-                }
-                
-                // Load link account session. Continue without Link if it errors.
-                let linkAccount = try? await _lookupLinkAccount(intent: intent, configuration: configuration)
-                LinkAccountContext.shared.account = linkAccount
-                
-                // Filter out payment methods that the PI/SI or PaymentSheet doesn't support
-                // TODO: We're fetching the customer's saved card and us_bank_account PMs, and then filtering - this is backwards!
-                let filteredSavedPaymentMethods = try await savedPaymentMethods
-                    .filter { intent.recommendedPaymentMethodTypes.contains($0.type) }
-                    .filter {
-                        $0.paymentSheetPaymentMethodType().supportsSavedPaymentMethod(
-                            configuration: configuration,
-                            intent: intent
-                        )
-                    }
-
-                // Warn the merchant if we see unactivated payment method types in the Intent
-                warnUnactivatedIfNeeded(unactivatedPaymentMethodTypes: intent.unactivatedPaymentMethodTypes)
-                // Ensure that there's at least 1 payment method type available for the intent and configuration.
-                let paymentMethodTypes = PaymentMethodType.filteredPaymentMethodTypes(from: intent, configuration: configuration)
-                guard !paymentMethodTypes.isEmpty else {
-                    completion(.failure(PaymentSheetError.noPaymentMethodTypesAvailable(intentPaymentMethods: intent.recommendedPaymentMethodTypes)))
-                    return
-                }
-                
-                completion(
-                    .success(
-                        intent: intent,
-                        savedPaymentMethods: filteredSavedPaymentMethods,
-                        isLinkEnabled: intent.supportsLink
-                    )
-                )
-            } catch {
-                completion(.failure(error))
-            }
-        }
-    }
-    
-    /// Loads miscellaneous singletons
-    static func loadMiscellaneousSingletons() async -> Bool {
-        await withCheckedContinuation { continuation in
-            Task {
-                AddressSpecProvider.shared.loadAddressSpecs {
-                    // Load form specs
-                    FormSpecProvider.shared.load { _ in
-                        // Load BSB data
-                        BSBNumberProvider.shared.loadBSBData {
-                            continuation.resume(returning: true)
-                        }
-                    }
-                }
-            }
-        }
-    }
-    
-    static func lookupLinkAccount(
-        intent: Intent,
-        configuration: Configuration
-    ) -> Promise<PaymentSheetLinkAccount?> {
-        // Only lookup the consumer account if Link is supported
-        guard intent.supportsLink else {
-            return .init(value: nil)
-        }
-
-        let promise = Promise<PaymentSheetLinkAccount?>()
-
-        let linkAccountService = LinkAccountService(apiClient: configuration.apiClient)
-
-        let consumerSessionLookupBlock: (String?) -> Void = { email in
-            if let email = email, linkAccountService.hasEmailLoggedOut(email: email) {
-                promise.resolve(with: nil)
-                return
-            }
-
-            linkAccountService.lookupAccount(withEmail: email) { result in
-                switch result {
-                case .success(let linkAccount):
-                    promise.resolve(with: linkAccount)
-                case .failure(let error):
-                    promise.reject(with: error)
-                }
-            }
-        }
-
-        if linkAccountService.hasSessionCookie {
-            consumerSessionLookupBlock(nil)
-        } else if let email = linkAccountService.getLastSignUpEmail() {
-            consumerSessionLookupBlock(email)
-        } else if let email = configuration.defaultBillingDetails.email {
-            consumerSessionLookupBlock(email)
-        } else if let customerID = configuration.customer?.id,
-            let ephemeralKey = configuration.customer?.ephemeralKeySecret
-        {
-            configuration.apiClient.retrieveCustomer(customerID, using: ephemeralKey) { customer, _ in
-                // If there's an error in this call we can just ignore it
-                consumerSessionLookupBlock(customer?.email)
-            }
-        } else {
-            promise.resolve(with: nil)
-        }
-
-        return promise
-    }
-
-    static func _lookupLinkAccount(
-        intent: Intent,
-        configuration: Configuration
-    ) async throws -> PaymentSheetLinkAccount? {
-        // Only lookup the consumer account if Link is supported
-        guard intent.supportsLink else {
-            return nil
-        }
-
-        let linkAccountService = LinkAccountService(apiClient: configuration.apiClient)
-        func lookUpConsumerSession(email: String?) async throws -> PaymentSheetLinkAccount? {
-            if let email = email, linkAccountService.hasEmailLoggedOut(email: email) {
-                return nil
-            }
-
-            return try await withCheckedThrowingContinuation { continuation in
-                linkAccountService.lookupAccount(withEmail: email) { result in
-                    switch result {
-                    case .success(let linkAccount):
-                        continuation.resume(with: .success(linkAccount))
-                    case .failure(let error):
-                        continuation.resume(throwing: error)
-                    }
-                }
-            }
-        }
-
-        if linkAccountService.hasSessionCookie {
-            return try await lookUpConsumerSession(email: nil)
-        } else if let email = linkAccountService.getLastSignUpEmail() {
-            return try await lookUpConsumerSession(email: email)
-        } else if let email = configuration.defaultBillingDetails.email {
-            return try await lookUpConsumerSession(email: email)
-        } else if let customerID = configuration.customer?.id,
-            let ephemeralKey = configuration.customer?.ephemeralKeySecret
-        {
-            let customer = try await configuration.apiClient.retrieveCustomer(customerID, using: ephemeralKey)
-            // If there's an error in this call we can just ignore it
-            return try await lookUpConsumerSession(email: customer.email)
-        } else {
-            return nil
-        }
-    }
-    
-    static func fetchIntent(mode: InitializationMode, configuration: Configuration) async throws -> Intent {
-        switch mode {
-        case .paymentIntentClientSecret(let clientSecret):
-            let paymentIntent: STPPaymentIntent
-            do {
-                paymentIntent = try await configuration.apiClient.retrievePaymentIntentWithPreferences(withClientSecret: clientSecret)
-            } catch {
-                // Fallback to regular retrieve PI when retrieve PI with preferences fails
-                paymentIntent = try await configuration.apiClient.retrievePaymentIntent(clientSecret: clientSecret)
-            }
-            guard ![.succeeded, .canceled, .requiresCapture].contains(paymentIntent.status) else {
-                // Error if the PaymentIntent is in a terminal state
-                let message = "PaymentSheet received a PaymentIntent in a terminal state: \(paymentIntent.status)"
-                throw PaymentSheetError.unknown(debugDescription: message)
-            }
-            return .paymentIntent(paymentIntent)
-        case .setupIntentClientSecret(let clientSecret):
-            let setupIntent: STPSetupIntent
-            do {
-                setupIntent = try await configuration.apiClient.retrieveSetupIntentWithPreferences(withClientSecret: clientSecret)
-            } catch {
-                // Fallback to regular retrieve SI when retrieve SI with preferences fails
-                setupIntent = try await configuration.apiClient.retrieveSetupIntent(clientSecret: clientSecret)
-            }
-            guard ![.succeeded, .canceled].contains(setupIntent.status) else {
-                // Error if the SetupIntent is in a terminal state
-                let message = "PaymentSheet received a SetupIntent in a terminal state: \(setupIntent.status)"
-                throw PaymentSheetError.unknown(debugDescription: message)
-            }
-            return .setupIntent(setupIntent)
-
-        case .deferredIntent(let intentConfig):
-            let elementsSession = try await configuration.apiClient.retrieveElementsSession(withIntentConfig: intentConfig)
-            return .deferredIntent(elementsSession: elementsSession, intentConfig: intentConfig)
-        }
-    }
-    
-    static func fetchSavedPaymentMethods(configuration: Configuration) async throws -> [STPPaymentMethod] {
-        let savedPaymentMethodTypes: [STPPaymentMethodType] = [.card, .USBankAccount]  // hardcoded for now
-        guard let customerID = configuration.customer?.id, let ephemeralKey = configuration.customer?.ephemeralKeySecret else {
-            return []
-        }
-        return try await withCheckedThrowingContinuation { continuation in
-            configuration.apiClient.listPaymentMethods(
-                forCustomer: customerID,
-                using: ephemeralKey,
-                types: savedPaymentMethodTypes
-            ) { paymentMethods, error in
-                guard let paymentMethods = paymentMethods, error == nil else {
-                    let error = error ?? PaymentSheetError.unknown(debugDescription: "Failed to retrieve PaymentMethods for the customer")
-                    continuation.resume(throwing: error)
-                    return
-                }
-                continuation.resume(returning: paymentMethods)
-            }
-        }
-    }
-
     // MARK: - Helper methods
-
-    private static func warnUnactivatedIfNeeded(unactivatedPaymentMethodTypes: [STPPaymentMethodType]) {
-        guard !unactivatedPaymentMethodTypes.isEmpty else { return }
-
-        let message = """
-            [Stripe SDK] Warning: Your Intent contains the following payment method types which are activated for test mode but not activated for live mode: \(unactivatedPaymentMethodTypes.map({ $0.displayName }).joined(separator: ",")). These payment method types will not be displayed in live mode until they are activated. To activate these payment method types visit your Stripe dashboard.
-            More information: https://support.stripe.com/questions/activate-a-new-payment-method
-            """
-        print(message)
-    }
 
     static func makeShippingParams(for paymentIntent: STPPaymentIntent, configuration: PaymentSheet.Configuration)
         -> STPPaymentIntentShippingDetailsParams?
